Extract common test setup
I had a brief play with making DataSet have a pointer to Storage and
DataSetMetaData. I abandoned that idea since it seemed to read worse,
with lots of pointer dereferencing and having to copy into temp
variables when using range to avoid all spawned goroutines pointing to
the last value of the pointer.

eg

```
for _, metaData := range datasets {
	go checkFreshness(dataset.DataSet{DataSetStorage, metaData}, failing, wg)
}
```

became

```
for _, metaData := range datasets {
    copy := metaData
	go checkFreshness(dataset.DataSet{&DataSetStorage, &copy}, failing, wg)
}
```

I spent a bit of time debugging this. So going for the similar version,
which doesn’t contain such quirks.