Implement JSON Schema validation
We are using github.com/xeipuuv/gojsonschema

Currently this implementation does the MAY aspect of the spec, and
doesn't validate formats.

See http://tools.ietf.org/html/draft-fge-json-schema-validation-00#section-7.2

We validate it separately ourselves for _timestamp. If other schemas are
defined with "format": "date-time", we may want to add those to our own validation.

Or we MAY want to implement format support ourselves in the go
implementation.