Add coverage statistics
Want to add support for looking at code coverage.

At the moment it’s a little suspect. My impression (from writing the
code) is that we have very good (~85%) coverage for the config_api
package, but it’s currently listed at 0.0% coverage of statements.

I am betting on the tools improving and this will be more useful. When
that happens, we should just be able to upgrade and see it working.

*UPDATE* actually, Travis reports it as 84.4% coverage, but running
locally gives 0% coverage for pkg/config_api. Not sure why?