Build on 1.4 and 1.5
Enforce an order on the DataSetStatus array to avoid test failures
caused by concurrent code returning different orders.

This is only needed for testing, so the other option would be to rewrite
the test to be more forgiving of the order of an array of ErrorInfo. Or
return a Set rather than an array.